display data list game to view

+++ app/src/main/java/com/dicoding/submission_capstone/core/data/source/remote/response/games/GameResponse.kt
@@ -21,7 +21,7 @@ data class GameResponse(
     @SerializedName("playtime")
     var playtime: Int?,
     @SerializedName("rating")
-    var rating: Int?,
+    var rating: Double?,
     @SerializedName("rating_top")
     var ratingTop: Int?,
     @SerializedName("ratings_count")
